@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+3
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -34,7 +34,6 @@ dependencies:
   firebase_database: ^3.0.9
   firebase_storage: ^3.0.8
 
-  keyboard_visibility: ^0.5.6
   image: ^2.1.0
   image_picker: ^0.5.3+2
   image_cropper: ^1.0.2
@@ -45,13 +44,9 @@ dependencies:
 
   flutter_staggered_grid_view: ^0.3.0
 
-  scrolling_page_indicator: ^0.1.2
-
-  flutter_swiper: ^1.1.6
   preload_page_view: ^0.1.4
 
-  smart_text_view: ^0.1.0
-
+  #for dm screen
   flutter_slidable: ^0.5.3
 
   bubble: ^1.1.8+2
@@ -59,13 +54,9 @@ dependencies:
   line_icons: ^0.2.0
   material_design_icons_flutter: ^3.2.3595
   flutter_vector_icons: ^0.1.0
-  bubble_tab_indicator: ^0.1.4
   extended_nested_scroll_view: ^0.3.4
   flutter_spinkit: ^3.1.0
-  gradient_widgets: ^0.4.1
-  auto_size_text: ^1.1.2
 
-  flutter_easyrefresh: ^1.2.7
   # state management
   get_it: ^3.0.1
   provider: ^3.0.0+1
