@@ -21,6 +21,7 @@ class RefreshListView extends StatefulWidget {
 
 class _RefreshListViewState extends State<RefreshListView> {
   bool isLoadingMore = false;
+  ScrollController controller;
 
   _loadMore() async {
     if (widget.onLoadMore != null) {
@@ -35,56 +36,54 @@ class _RefreshListViewState extends State<RefreshListView> {
 
   @override
   void initState() {
+    controller = widget.controller ?? ScrollController();
+    controller.addListener(() {
+      if (!isLoadingMore &&
+          controller.position.pixels > controller.position.maxScrollExtent) {
+        _loadMore();
+        isLoadingMore = true;
+
+        if (mounted) setState(() {});
+
+        Future.delayed(Duration(seconds: 15))
+          ..whenComplete(() {
+            isLoadingMore = false;
+
+            if (mounted) setState(() {});
+          });
+      }
+    });
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return NotificationListener<ScrollNotification>(
-      onNotification: (ScrollNotification scrollInfo) {
-        if (!isLoadingMore &&
-            scrollInfo.metrics.pixels > scrollInfo.metrics.maxScrollExtent) {
-          _loadMore();
-          isLoadingMore = true;
-
-          if (mounted) setState(() {});
-
-          Future.delayed(Duration(seconds: 5))
-            ..whenComplete(() {
-              isLoadingMore = false;
-
-              if (mounted) setState(() {});
-            });
-        }
-        return true;
-      },
-      child: CustomScrollView(
-        controller: widget.controller,
-        physics: AlwaysScrollableScrollPhysics(),
-        slivers: <Widget>[
-          CupertinoSliverRefreshControl(
-            refreshIndicatorExtent: 80,
-            refreshTriggerPullDistance: 120,
-            onRefresh: widget.onRefresh,
-            builder: (context, mode, _, __, ___) => Padding(
-              padding: const EdgeInsets.all(20.0),
-              child: CupertinoActivityIndicator(
-                radius: 12,
-              ),
+    return CustomScrollView(
+      controller: widget.controller,
+      physics: AlwaysScrollableScrollPhysics(),
+      slivers: <Widget>[
+        CupertinoSliverRefreshControl(
+          refreshIndicatorExtent: 80,
+          refreshTriggerPullDistance: 120,
+          onRefresh: widget.onRefresh,
+          builder: (context, mode, _, __, ___) => Padding(
+            padding: const EdgeInsets.all(20.0),
+            child: CupertinoActivityIndicator(
+              radius: 12,
             ),
           ),
-          ...widget.children.map((child) => SliverToBoxAdapter(
-                child: child,
-              )),
-          if (isLoadingMore)
-            SliverToBoxAdapter(
-              child: Padding(
-                padding: const EdgeInsets.all(10.0),
-                child: CupertinoActivityIndicator(),
-              ),
-            )
-        ],
-      ),
+        ),
+        ...widget.children.map((child) => SliverToBoxAdapter(
+              child: child,
+            )),
+        if (isLoadingMore)
+          SliverToBoxAdapter(
+            child: Padding(
+              padding: const EdgeInsets.all(10.0),
+              child: CupertinoActivityIndicator(),
+            ),
+          )
+      ],
     );
   }
 }
